solved issue with dist2receiver in carve

+++ include/fastscapelib/basin_graph.hpp
@@ -663,7 +663,7 @@ auto get_d8_distances_sep(double dx, double dy) -> std::array<double, 9>
 {
     std::array<double, 9> d8_dists;
 
-    for(size_t k=0; k<9; ++k)
+    for(int k=0; k<9; ++k)
     {
         double d8_dx = dx * double(k % 3 -1);
         double d8_dy = dy * double(k / 3 -1);
@@ -771,13 +771,18 @@ void BasinGraph<Basin_T, Node_T, Elevation_T>::update_pits_receivers_carve(Rcv_X
         Elevation_T previous_dist = dist2receivers[cur_node];
 
         receivers(cur_node) = link.nodes[OUTFLOW];
+        //std::cerr << "+ [" << cur_node << "]" << dist2receivers(cur_node);
         dist2receivers(cur_node) = d8_distances[detail::get_d8_distance_id(cur_node, link.nodes[OUTFLOW], ncols)];
+        //std::cerr << "->" << dist2receivers(cur_node)<< std::endl;
 
         //std::cout << "Pass " << cur_node << " -> " << link.nodes[OUTFLOW] << std::endl;
 
         while( cur_node != outlet_inflow)
         {
+            //std::cerr << "  [" << next_node << "]" << dist2receivers(next_node);
             std::swap(dist2receivers(next_node), previous_dist);
+            //std::cerr << "->" << dist2receivers(cur_node)<< std::endl;
+
 
             Node_T rcv_next_node = receivers(next_node);
             receivers(next_node) = cur_node;
